+++ R/RcppExports.R
@@ -13,6 +13,10 @@ mh_binom_mv <- function(qt, p, proposal, k, n, mean, Q, use_norm, norm, acceptan
     .Call('_gibbs_utils_mh_binom_mv', PACKAGE = 'gibbs.utils', qt, p, proposal, k, n, mean, Q, use_norm, norm, acceptance)
 }
 
+mh_multinom_mv <- function(qt, p_ij, proposal, z_ij, which_i, is_ref, k_ij, n_ij, mean, Q, use_norm, norm, diag, acceptance) {
+    .Call('_gibbs_utils_mh_multinom_mv', PACKAGE = 'gibbs.utils', qt, p_ij, proposal, z_ij, which_i, is_ref, k_ij, n_ij, mean, Q, use_norm, norm, diag, acceptance)
+}
+
 mh_pois <- function(method, L, proposal, k, k_na, mean, precision, trunc_at, lower, acceptance) {
     .Call('_gibbs_utils_mh_pois', PACKAGE = 'gibbs.utils', method, L, proposal, k, k_na, mean, precision, trunc_at, lower, acceptance)
 }
@@ -37,8 +41,8 @@ one_multinom_slice <- function(p_ij, sum_exp_p, z_ij, k, n, mean, precision, w, 
     .Call('_gibbs_utils_one_multinom_slice', PACKAGE = 'gibbs.utils', p_ij, sum_exp_p, z_ij, k, n, mean, precision, w, nexpand, ncontract)
 }
 
-slice_sample_multinom_mv <- function(p_ij, z_ij, which_i, is_ref, k_ij, n_ij, mean, Q, diag, w, nexpand, ncontract) {
-    .Call('_gibbs_utils_slice_sample_multinom_mv', PACKAGE = 'gibbs.utils', p_ij, z_ij, which_i, is_ref, k_ij, n_ij, mean, Q, diag, w, nexpand, ncontract)
+slice_sample_multinom_mv <- function(p_ij, z_ij, which_i, is_ref, k_ij, n_ij, mean, Q, diag, use_norm, norm, w, nexpand, ncontract) {
+    .Call('_gibbs_utils_slice_sample_multinom_mv', PACKAGE = 'gibbs.utils', p_ij, z_ij, which_i, is_ref, k_ij, n_ij, mean, Q, diag, use_norm, norm, w, nexpand, ncontract)
 }
 
 one_pois_slice <- function(L, k, mean, precision, trunc_at, lower, w, nexpand, ncontract) {