+++ src/RcppExports.cpp
@@ -64,6 +64,30 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// mh_multinom_mv
+NumericVector mh_multinom_mv(bool qt, NumericMatrix p_ij, NumericMatrix proposal, LogicalMatrix z_ij, IntegerVector which_i, LogicalVector is_ref, NumericMatrix k_ij, NumericMatrix n_ij, NumericMatrix mean, NumericMatrix Q, LogicalVector use_norm, NumericMatrix norm, bool diag, int acceptance);
+RcppExport SEXP _gibbs_utils_mh_multinom_mv(SEXP qtSEXP, SEXP p_ijSEXP, SEXP proposalSEXP, SEXP z_ijSEXP, SEXP which_iSEXP, SEXP is_refSEXP, SEXP k_ijSEXP, SEXP n_ijSEXP, SEXP meanSEXP, SEXP QSEXP, SEXP use_normSEXP, SEXP normSEXP, SEXP diagSEXP, SEXP acceptanceSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< bool >::type qt(qtSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type p_ij(p_ijSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type proposal(proposalSEXP);
+    Rcpp::traits::input_parameter< LogicalMatrix >::type z_ij(z_ijSEXP);
+    Rcpp::traits::input_parameter< IntegerVector >::type which_i(which_iSEXP);
+    Rcpp::traits::input_parameter< LogicalVector >::type is_ref(is_refSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type k_ij(k_ijSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type n_ij(n_ijSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type mean(meanSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type Q(QSEXP);
+    Rcpp::traits::input_parameter< LogicalVector >::type use_norm(use_normSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type norm(normSEXP);
+    Rcpp::traits::input_parameter< bool >::type diag(diagSEXP);
+    Rcpp::traits::input_parameter< int >::type acceptance(acceptanceSEXP);
+    rcpp_result_gen = Rcpp::wrap(mh_multinom_mv(qt, p_ij, proposal, z_ij, which_i, is_ref, k_ij, n_ij, mean, Q, use_norm, norm, diag, acceptance));
+    return rcpp_result_gen;
+END_RCPP
+}
 // mh_pois
 NumericVector mh_pois(int method, NumericVector L, NumericVector proposal, NumericVector k, LogicalVector k_na, NumericVector mean, NumericVector precision, NumericVector trunc_at, LogicalVector lower, int acceptance);
 RcppExport SEXP _gibbs_utils_mh_pois(SEXP methodSEXP, SEXP LSEXP, SEXP proposalSEXP, SEXP kSEXP, SEXP k_naSEXP, SEXP meanSEXP, SEXP precisionSEXP, SEXP trunc_atSEXP, SEXP lowerSEXP, SEXP acceptanceSEXP) {
@@ -183,8 +207,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // slice_sample_multinom_mv
-NumericMatrix slice_sample_multinom_mv(NumericMatrix p_ij, LogicalMatrix z_ij, IntegerVector which_i, LogicalVector is_ref, NumericMatrix k_ij, NumericMatrix n_ij, NumericMatrix mean, NumericMatrix Q, bool diag, double w, int nexpand, int ncontract);
-RcppExport SEXP _gibbs_utils_slice_sample_multinom_mv(SEXP p_ijSEXP, SEXP z_ijSEXP, SEXP which_iSEXP, SEXP is_refSEXP, SEXP k_ijSEXP, SEXP n_ijSEXP, SEXP meanSEXP, SEXP QSEXP, SEXP diagSEXP, SEXP wSEXP, SEXP nexpandSEXP, SEXP ncontractSEXP) {
+NumericMatrix slice_sample_multinom_mv(NumericMatrix p_ij, LogicalMatrix z_ij, IntegerVector which_i, LogicalVector is_ref, NumericMatrix k_ij, NumericMatrix n_ij, NumericMatrix mean, NumericMatrix Q, bool diag, LogicalVector use_norm, NumericMatrix norm, double w, int nexpand, int ncontract);
+RcppExport SEXP _gibbs_utils_slice_sample_multinom_mv(SEXP p_ijSEXP, SEXP z_ijSEXP, SEXP which_iSEXP, SEXP is_refSEXP, SEXP k_ijSEXP, SEXP n_ijSEXP, SEXP meanSEXP, SEXP QSEXP, SEXP diagSEXP, SEXP use_normSEXP, SEXP normSEXP, SEXP wSEXP, SEXP nexpandSEXP, SEXP ncontractSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -197,10 +221,12 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< NumericMatrix >::type mean(meanSEXP);
     Rcpp::traits::input_parameter< NumericMatrix >::type Q(QSEXP);
     Rcpp::traits::input_parameter< bool >::type diag(diagSEXP);
+    Rcpp::traits::input_parameter< LogicalVector >::type use_norm(use_normSEXP);
+    Rcpp::traits::input_parameter< NumericMatrix >::type norm(normSEXP);
     Rcpp::traits::input_parameter< double >::type w(wSEXP);
     Rcpp::traits::input_parameter< int >::type nexpand(nexpandSEXP);
     Rcpp::traits::input_parameter< int >::type ncontract(ncontractSEXP);
-    rcpp_result_gen = Rcpp::wrap(slice_sample_multinom_mv(p_ij, z_ij, which_i, is_ref, k_ij, n_ij, mean, Q, diag, w, nexpand, ncontract));
+    rcpp_result_gen = Rcpp::wrap(slice_sample_multinom_mv(p_ij, z_ij, which_i, is_ref, k_ij, n_ij, mean, Q, diag, use_norm, norm, w, nexpand, ncontract));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -293,13 +319,14 @@ static const R_CallMethodDef CallEntries[] = {
     {"_gibbs_utils_impute_conj_mvnorm_mu_cpp", (DL_FUNC) &_gibbs_utils_impute_conj_mvnorm_mu_cpp, 6},
     {"_gibbs_utils_mh_binom", (DL_FUNC) &_gibbs_utils_mh_binom, 8},
     {"_gibbs_utils_mh_binom_mv", (DL_FUNC) &_gibbs_utils_mh_binom_mv, 10},
+    {"_gibbs_utils_mh_multinom_mv", (DL_FUNC) &_gibbs_utils_mh_multinom_mv, 14},
     {"_gibbs_utils_mh_pois", (DL_FUNC) &_gibbs_utils_mh_pois, 10},
     {"_gibbs_utils_mh_pois_mv", (DL_FUNC) &_gibbs_utils_mh_pois_mv, 12},
     {"_gibbs_utils_one_binom_slice", (DL_FUNC) &_gibbs_utils_one_binom_slice, 8},
     {"_gibbs_utils_slice_sample_binom", (DL_FUNC) &_gibbs_utils_slice_sample_binom, 8},
     {"_gibbs_utils_slice_sample_binom_mv", (DL_FUNC) &_gibbs_utils_slice_sample_binom_mv, 10},
     {"_gibbs_utils_one_multinom_slice", (DL_FUNC) &_gibbs_utils_one_multinom_slice, 10},
-    {"_gibbs_utils_slice_sample_multinom_mv", (DL_FUNC) &_gibbs_utils_slice_sample_multinom_mv, 12},
+    {"_gibbs_utils_slice_sample_multinom_mv", (DL_FUNC) &_gibbs_utils_slice_sample_multinom_mv, 14},
     {"_gibbs_utils_one_pois_slice", (DL_FUNC) &_gibbs_utils_one_pois_slice, 9},
     {"_gibbs_utils_slice_sample_pois", (DL_FUNC) &_gibbs_utils_slice_sample_pois, 10},
     {"_gibbs_utils_slice_sample_pois_mv", (DL_FUNC) &_gibbs_utils_slice_sample_pois_mv, 12},